backend/requirements.txt
fastapi
uvicorn
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv
passlib==1.7.4
bcrypt==3.2.2
python-multipart
python-jose[cryptography]
qrcode
pillow
resend
email-validator